+++ lib/providers/firebase-admin.js
@@ -4,6 +4,7 @@ import {initializeApp} from 'firebase-admin/app'
 import {getAuth} from 'firebase-admin/auth'
 import {getDatabase} from 'firebase-admin/database'
 import config from 'lib/config'
+import {toArray} from 'lib/providers/firebase-rest'
 
 const {
 	FIREBASE_DATABASE_URL,
@@ -27,7 +28,6 @@ initializeApp({
 		privateKey: FIREBASE_SERVICE_ACCOUNT_PRIVATE_KEY,
 	}),
 })
-// }
 
 async function getUser(uid) {
 	const db = getDatabase()
@@ -38,13 +38,13 @@ async function getUser(uid) {
 async function getChannel(id) {
 	const db = getDatabase()
 	const snapshot = await db.ref(`channels/${id}`).once('value')
-	return snapshot.val()
+	return serializeChannel(snapshot.val())
 }
 
 async function getChannelTracks(channelId) {
 	const db = getDatabase()
 	const snapshot = await db.ref(`tracks`).orderByChild('channel').equalTo(channelId).once('value')
-	return snapshot.val()
+	return serializeTracks(snapshot.val())
 }
 
 async function getUserExport(uid) {
@@ -55,5 +55,36 @@ async function getUserExport(uid) {
 	return {user, channel, tracks}
 }
 
+
+/* serialize functions all expect a `snapshot.val()` firebase-admin value */
+function serializeChannel(channel) {
+	if (!channel) return null
+	channel.created = toTimestamp(channel.created)
+	channel.updated = channel.updated ? toTimestamp(channel.updated) : channel.created
+	return channel
+}
+
+function serializeTracks(tracks) {
+	if (!tracks) return null
+	tracks = toArray(tracks)
+	if (tracks.length > 0) {
+		tracks = tracks.map((track) => {
+			if (!track.title) track.title = 'Untitled'
+			track.created = toTimestamp(track.created)
+			return track
+		})
+	}
+	return tracks
+}
+
+/*
+	Converts the Firebase timestamps to what Postgres wants:
+	new Date("1411213745028").toISOString() ==> "2014-09-20T11:49:05.028Z"
+*/
+function toTimestamp(timestamp) {
+	return new Date(Number(timestamp)).toISOString()
+}
+
+
 export default adminClient
-export {getAuth, getUser, getUserExport}
+export {getAuth, getUserExport}
