+++ lib/providers/firebase-rest.js
@@ -27,11 +27,11 @@ async function fetchAndParse(url) {
 	return data
 }
 
-function toObject(obj, id) {
+export function toObject(obj, id) {
 	return Object.assign(obj, {id})
 }
 
-function toArray(data) {
+export function toArray(data) {
 	return Object.keys(data).map((id) => toObject(data[id], id))
 }
